Move tests for the zFCP module
Move the unit tests for the zFCP module to a new file.
